backend/requirements.txt
# FastAPI & Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic[email]==2.5.3

# WebSockets
python-socketio==5.11.0
websockets==12.0

# AI Integration
google-generativeai==0.3.2

# HTTP Client (for external APIs)
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Date/Time
python-dateutil==2.8.2

# Background Tasks
celery==5.3.6
celery[redis]==5.3.6

# Rate Limiting
slowapi==0.1.9

# CORS
python-cors==1.0.0

# Logging
structlog==24.1.0

# Testing (production)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
